Count and display % of completion of the project
Create new functions to counting percent of completion of the project.
This information is showing in card of project in main view.
The percentage is calculated on the basis of during time tasks, amount
of tasks and weight of tasks.
We create new action addTaskToProject, beacuse we haven't had it yet
and it was needed for the calculations